Refactor allowlist editor
