+++ packages/browser-wallet/src/popup/pages/ConnectAccountsRequest/ConnectAccountsRequest.scss
@@ -7,51 +7,3 @@
     border: rem(1px) solid #dae4ee;
     border-radius: rem(10px) rem(10px) 0 0;
 }
-
-.connect-accounts-request-accounts {
-    padding-left: rem(10px);
-    padding-right: rem(10px);
-    width: 100%;
-    min-height: rem(175px);
-    margin-left: auto;
-    margin-right: auto;
-    margin-bottom: rem(15px);
-    background: $color-input-bg;
-    border: rem(1px) solid #dae4ee;
-    border-top: none;
-    border-radius: 0 0 rem(10px) rem(10px);
-    scrollbar-width: thin;
-    overflow-y: auto;
-
-    &__account-item {
-        padding-top: rem(5px);
-        padding-bottom: rem(5px);
-        font-weight: $font-weight-light;
-
-        &:not(:last-child) {
-            border-bottom: rem(1px) dashed #dae4ee;
-        }
-
-        &__primary {
-            position: relative;
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-        }
-
-        &__check-box {
-            position: absolute;
-            right: rem(0);
-            bottom: rem(-12.5px);
-            width: rem(15px);
-
-            input {
-                margin: 0;
-            }
-        }
-
-        &__secondary {
-            font-size: rem(8px);
-        }
-    }
-}
